Rename src/chronograf to src/data_explorer
