+++ ui/src/data_explorer/actions/view/index.js
