+++ ui/src/data_explorer/components/GroupByTimeDropdown.js
