+++ ui/src/data_explorer/components/Panel.js
