+++ ui/src/data_explorer/components/PanelBuilder.js
