+++ ui/src/data_explorer/components/PanelList.js
