+++ ui/src/data_explorer/components/QueryEditor.js
