+++ ui/src/data_explorer/components/QueryTabItem.js
