+++ ui/src/data_explorer/components/RenamePanelModal.js
