+++ ui/src/data_explorer/data/groupByTimes.hson
