+++ ui/src/data_explorer/reducers/activeExplorer.js
