+++ ui/src/data_explorer/reducers/explorers.js
