+++ ui/src/data_explorer/reducers/index.js
