+++ ui/src/data_explorer/reducers/panels.js
