+++ ui/src/data_explorer/reducers/queryConfigs.js
