+++ ui/src/data_explorer/reducers/timeRange.js
