+++ ui/src/kapacitor/components/KapacitorRule.js
@@ -5,7 +5,7 @@ import RuleHeader from 'src/kapacitor/components/RuleHeader';
 import RuleGraph from 'src/kapacitor/components/RuleGraph';
 import RuleMessage from 'src/kapacitor/components/RuleMessage';
 import {createRule, editRule} from 'src/kapacitor/apis';
-import selectStatement from '../../chronograf/utils/influxql/select';
+import selectStatement from '../../data_explorer/utils/influxql/select';
 import timeRanges from 'hson!../../shared/data/timeRanges.hson';
 
 export const KapacitorRule = React.createClass({
